fixed draw sphere missing an edge, fixed billboard rendering pos, added clone to material, reset renderbuffer in framebuffer

+++ Engine/src/BHive/renderers/render_passes/OutlineRenderPass.cpp
@@ -35,7 +35,7 @@ namespace BHive
 		FramebufferSpecification specs{};
 		specs.Width = mSize.x;
 		specs.Height = mSize.y;
-		specs.Attachments.attach(FTextureCreateInfo{.InternalFormat = EFormat::RGBA8, .WrapMode = EWrapMode::CLAMP_TO_EDGE}, ETextureType::TEXTURE_2D);
+		specs.Attachments.attach(FTextureCreateInfo{.InternalFormat = EFormat::RGBA32F, .WrapMode = EWrapMode::CLAMP_TO_EDGE}, ETextureType::TEXTURE_2D);
 		specs.Attachments.attach(FRenderbufferTexture{.Format = EFormat::DEPTH24_STENCIL8});
 		mFrambuffer = CreateRef<Framebuffer>(specs);
 	}
@@ -56,7 +56,7 @@ namespace BHive
 		mBoxBlurShader = ShaderManager::Get().Load("BoxBlur.glsl");
 
 		FTextureCreateInfo create_info_lut{};
-		create_info_lut.InternalFormat = EFormat::RGBA8;
+		create_info_lut.InternalFormat = EFormat::RGBA32F;
 		create_info_lut.WrapMode = EWrapMode::REPEAT;
 		create_info_lut.MinFilter = EMinFilter::LINEAR;
 
@@ -77,7 +77,7 @@ namespace BHive
 
 		FTextureCreateInfo create_info{};
 		create_info.WrapMode = EWrapMode::CLAMP_TO_EDGE;
-		create_info.InternalFormat = EFormat::RGBA8;
+		create_info.InternalFormat = EFormat::RGBA32F;
 
 		mOutputTexture = CreateRef<Texture2D>(size.x, size.y, create_info);
 		mOutlineOutput = CreateRef<Texture2D>(size.x, size.y, create_info);
